server/pyproject.toml
[tool.poetry]
name = "artist-recommendations"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "artist_recommendations"}]

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^3.0.1"
requests = "^2.31.0"
flask-cors = "^4.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
